Apply project config on runmap and and set lua flag for lua builds
